refactor: move diff index to output contracts
